ds2.11a22

+++ lib/cmds/creators/invis.c
@@ -8,7 +8,7 @@ inherit LIB_DAEMON;
 
 varargs int cmd()
 {
-    if((int)this_player()->GetInvis()) {
+    if(this_player()->GetInvis()) {
         notify_fail("You are already invisible.\n");
         return 0;
     }
