+++ lib/cmds/players/date.c
@@ -16,7 +16,7 @@ mixed cmd(string timezone) {
     }
     if(!timezone || !valid_timezone(timezone)) timezone = local_time()[9];
 
-    offset = (int)TIME_D->GetOffset(timezone);
+    offset = TIME_D->GetOffset(timezone);
     offset += EXTRA_TIME_OFFSET;
     if(query_os_type() != "windows" ) 
         x = offset * 3600;
