+++ lib/cmds/players/equipment.c
@@ -6,7 +6,7 @@ inherit LIB_DAEMON;
 void eventInventory();
 
     mixed cmd(string args) {
-        if( (int)this_player()->GetInCombat() )
+        if( this_player()->GetInCombat() )
             this_player()->SetAttack(0, (: eventInventory :), ROUND_OTHER);
         else eventInventory();
         return 1;
