+++ lib/secure/cmds/admins/addguest.c
@@ -14,7 +14,7 @@ int cmd(string str) {
     mixed res;
     int i;
 
-    if(!(int)master()->valid_apply(({ PRIV_SECURE, PRIV_ASSIST }))){
+    if(!master()->valid_apply(({ PRIV_SECURE, PRIV_ASSIST }))){
         error("Illegal attempt to add a guest.");
     }
