+++ lib/secure/cmds/creators/stupidemote.c
@@ -15,7 +15,7 @@ mixed cmd(string args){
     string array rules = ({ "", "STR" });
     string array adverbs;
     string emote, xtra, prep;
-    if( !((int)master()->valid_apply(({ "ASSIST" }))) &&
+    if( !(master()->valid_apply(({ "ASSIST" }))) &&
             !member_group(this_player(), "EMOTES") ){
         write("You are not admin, nor a member of the EMOTES group.");
         return 1;
