+++ lib/secure/cmds/players/history.c
@@ -16,7 +16,7 @@ mixed cmd(string args) {
     if(player->GetForced()) return 0;
 
     if( x = to_int(args) ) {
-        if( (int)previous_object()->SetHistorySize(x) != x ) {
+        if( previous_object()->SetHistorySize(x) != x ) {
             if( x < MIN_HISTORY_SIZE )
                 return "History size must be at least " + MIN_HISTORY_SIZE + ".";
             else if( x > MAX_HISTORY_SIZE )
@@ -33,8 +33,8 @@ mixed cmd(string args) {
         mapping history;
         int cmd_num, i;
 
-        //history = (string array)previous_object()->GetHistoryList();
-        //cmd_num = (int)previous_object()->GetCommandNumber();
+        //history = previous_object()->GetHistoryList();
+        //cmd_num = previous_object()->GetCommandNumber();
         history = this_player()->GetCommandHist();
         cmd_num = sizeof(history);
         //x = (cmd_num - 2) % sizeof(history);
