+++ lib/secure/daemon/bboard.c
@@ -50,7 +50,7 @@ static private void restore_board() {
 static private int valid_access() {
     string str;
     if(this_player() && archp(this_player())) true();
-    else if(__Owner == PRIV_SECURE && !((int)master()->valid_apply(({}))))
+    else if(__Owner == PRIV_SECURE && !(master()->valid_apply(({}))))
         return 0;
     str = query_privs(previous_object(0));
     if(member_array(PRIV_SECURE, explode(str, ":")) != -1) return 1;
