+++ lib/secure/daemon/i3router/server.c
@@ -28,7 +28,7 @@ object ssocket = find_object(SSOCKET_D);
 static void validate(){
     if( previous_object() != cmd && previous_object() != rsocket &&
             previous_object() != this_object() && previous_object() != ssocket &&
-            !((int)master()->valid_apply(({ "ASSIST" }))) ){
+            !(master()->valid_apply(({ "ASSIST" }))) ){
         trr("SECURITY ALERT: validation failure in ROUTER_D.","red");
         error("Illegal attempt to access router daemon: "+get_stack()+
                 " "+identify(previous_object(-1)));
