+++ lib/secure/daemon/reaper.c
@@ -67,7 +67,7 @@ static void heart_beat(){
 }
 
 int eventDestruct(){
-    if( !((int)master()->valid_apply(({ "SECURE" }))) )
+    if( !(master()->valid_apply(({ "SECURE" }))) )
         error("Illegal attempt to destruct reaper: "+get_stack()+" "+identify(previous_object(-1)));
     return ::eventDestruct();
 }
