+++ lib/secure/sefun/local_time.c
@@ -77,7 +77,7 @@ string set_tz(string str){
         return "Invalid time zone.";
     }
     tz = str;
-    if( !((int)master()->valid_apply(({ "PRIV_ASSIST", "PRIV_SECURE" }))) ) 
+    if( !(master()->valid_apply(({ "PRIV_ASSIST", "PRIV_SECURE" }))) ) 
         error("Illegal attempt to modify timezone: "+get_stack()+" "+identify(previous_object(-1)));
     unguarded( (: write_file("/cfg/timezone.cfg",tz,1) :) );
     return "Mud time zone is now "+read_file("/cfg/timezone.cfg");
