+++ lib/verbs/items/dig.c
@@ -26,7 +26,7 @@ static void eventDig(object who, object tool, object what, object check) {
 varargs static void eventPrepareDig(object who, object tool, object what) {
     function f = (: eventDig($(who), $(tool), $(what), environment($(who))) :);
 
-    if( (int)this_player()->GetInCombat() ) {
+    if( this_player()->GetInCombat() ) {
         send_messages("start", "$agent_name $agent_verb to dig with " +
                 tool->GetShort() + ".", who, 0, environment(who));
         who->SetAttack(0, f, ROUND_OTHER);
