+++ lib/verbs/items/poison.c
@@ -29,7 +29,7 @@ static void create() {
 mixed can_poison_obj_with_obj(string verb) { return this_player()->CanManipulate(); }
 
     mixed do_poison_obj_with_obj(object target, object agent) {
-        if( (int)this_player()->GetInCombat() )
+        if( this_player()->GetInCombat() )
             this_player()->SetAttack(0, (: eventPoison, this_player(), target,
                         agent :), ROUND_OTHER);
         else eventPoison(this_player(), target, agent);
