+++ lib/verbs/players/teach.c
@@ -17,7 +17,7 @@ static void create()
 mixed can_teach_str_to_liv(string str, object ob)
 {
     int pos = this_player()->GetPosition();
-    if( (int)this_player()->GetParalyzed() ) {
+    if( this_player()->GetParalyzed() ) {
         return "You cannot move!";
     }
     if( pos == POSITION_SITTING || pos == POSITION_LYING &&
